refactor(webhook): split handlers.ts + handlers-app.ts into cohesive modules
Eliminate the last two webhook file-length exemptions by splitting along
responsibility seams (all resulting files <300 SLOC):

handlers.ts (462 → 124): dispatcher/webhookRoute only
- handlers-issues.ts (214): issues / issue_dependencies / sub_issues
- handlers-ref.ts (128): create / delete / pull_request / milestone

handlers-app.ts (363 → 171): install-lifecycle handlers only
- handlers-access.ts (171): repository / member / membership
- handlers-app-helpers.ts (28): shared payload-shape helpers

Pure moves, no logic change. Importers (handlers.ts dispatcher,
handlers.test.ts, handlers-app.test.ts) updated to direct sibling imports;
router.ts unaffected (webhookRoute stays in handlers.ts). CLAUDE.md Key
files table updated. Removes the handlers.ts + handlers-app.ts exemptions.
