+++ tools/file_exemptions.txt
@@ -5,8 +5,6 @@
 
 # Production sources above QG_FILE_MAX=300 (split deferred; cap prevents growth)
 worker/src/sync/sync.ts  # 1306 lines — #180 sync orchestrator; #216 structure-only sync path (+biome reflow)
-worker/src/webhook/handlers.ts  # 462 lines — #180 issue webhook dispatch + bootstrap sync trigger
-worker/src/webhook/handlers-app.ts  # 384 lines — #180 GitHub App lifecycle + install sender→user_installations link
 
 # Test files (colocated *.test.ts — exempt with local caps, not production code)
 worker/src/sync/sync.test.ts  # 1745 lines — #180 integration harness for runSync; #216 structure-only (+biome reflow)
